got posting channels to work and generate in the channel list.
also added a channel feature to redux store and added a channelSlice to handle the state of the channels.

modified some models to work properly with the new channel feature.

added a channelRoutes file to handle the routes for channels.

added a new css file for the right main container.

started working on right main container to display posts